refactor: move walls-and-chambers to an Oscar package extension
