+++ src/WallsAndChambers.jl
@@ -1,3 +1,15 @@
+# Walls-and-chambers and VGIT functionality relies on Oscar's polyhedral geometry
+# and therefore lives in the Oscar package extension (ext/QuiverToolsOscarExt.jl).
+# The functions below are only method stubs: the extension adds the real methods
+# once Oscar is loaded. Until then, calling any of them raises a clear error.
+#
+# To enable them, load Oscar alongside QuiverTools. Use `import Oscar` rather than
+# `using Oscar`: it activates the extension without bringing Oscar's exports into
+# scope, which would clash with QuiverTools names such as `index` and `todd_class`.
+#
+#     using QuiverTools
+#     import Oscar
+
 """
     is_special_subdimension_vector(Q::Quiver, e::AbstractVector{Int}, d::AbstractVector{Int})
 
@@ -5,6 +17,8 @@ Compute whether `e` is a special subdimension vector of `d` for `Q`.
 
 Special subdimension vectors are defined in
 [[Definition 6.1, MR5007902](https://mathscinet.ams.org/mathscinet-getitem?mr=5007902)].
+
+Requires Oscar: run `import Oscar` to enable this function.
 
 # Example
 
@@ -18,44 +32,14 @@ julia> is_special_subdimension_vector(Q, [1, 0, 0, 1], d)
 true
 ```
 """
-function is_special_subdimension_vector(
-  Q::Quiver, e::AbstractVector{Int}, d::AbstractVector{Int}
-)
-  e = QuiverTools.coerce_vector(e)
-  d = QuiverTools.coerce_vector(d)
+function is_special_subdimension_vector end
 
-  (all(ei -> ei == 0, e) || e == d) && return false
-  is_general_subdimension_vector(Q, e, d) && return false
-
-  eprimes = filter(
-    eprime -> !all(epi -> epi == 0, eprime),
-    all_general_subdimension_vectors(Q, e),
-  )
-
-  fprimes = filter(
-    fprime -> !all(fpi -> fpi == 0, fprime),
-    all_general_subdimension_vectors(Q, d - e),
-  )
-
-  cone = Oscar.polyhedron(
-    vcat(eprimes, [e + f for f in fprimes]),
-    zeros(Int, length(eprimes) + length(fprimes)),
-  )
-  cone = intersect(cone, sst(Q, d))
-
-  return !any(
-    issubset(cone, Oscar.polyhedron([eprime, -eprime], [0, 0]))
-    for eprime in eprimes
-  ) &&
-         !any(
-    issubset(cone, Oscar.polyhedron([e + f, - (e + f)], [0, 0]))
-    for f in fprimes if f != d - e # this would check if cone ⋐ d^⟂, which is true of course
-  )
-end
 """
     all_special_subdimension_vectors(Q::Quiver, d::AbstractVector{Int})
 
 Compute all the special subdimension vectors of `d` for the quiver `Q`.
+
+Requires Oscar: run `import Oscar` to enable this function.
 
 # Example
 
@@ -66,31 +50,7 @@ julia> length(all_special_subdimension_vectors(Q, d))
 7
 ```
 """
-@memoize Dict function all_special_subdimension_vectors(Q::Quiver, d::AbstractVector{Int})
-  d = QuiverTools.coerce_vector(d)
-  candidates = filter(
-    e -> !QuiverTools.is_general_subdimension_vector(Q, e, d),
-    QuiverTools.all_subdimension_vectors(d, nonzero=true, strict=true),
-  )
-  return filter(
-    e -> is_special_subdimension_vector(Q, e, d),
-    candidates,
-  )
-end
-
-"""
-    __helper_accelerate(P::Oscar.Polyhedron)
-
-Internal method.
-
-Convert the polyhedron `P` to be spanned by its rays.
-Only works for strongly convex polyhedra.
-"""
-function __helper_accelerate(P::Oscar.Polyhedron)
-  newP = Oscar.polyhedron(Oscar.positive_hull(Oscar.rays(P)))
-  newP != P && return P
-  return newP
-end
+function all_special_subdimension_vectors end
 
 """
     sst(Q, e)
@@ -99,28 +59,21 @@ Compute the semistable cone `sst(e)` for a given quiver `Q` and a vector `e`.
 This is the cone of all stability parameters for which semistable representations
 of the quiver `Q` with dimension vector `e` exist.
 
+Requires Oscar: run `import Oscar` to enable this function.
 
 # Example
 
 ```jldoctests
 julia> Q = Quiver("1-2,2-3,3-4,1-3,1-4"); d = [1, 1, 1, 1];
 
-julia> collect(QuiverTools.Oscar.rays(sst(Q, d)))
+julia> collect(Oscar.rays(sst(Q, d)))
 3-element Vector{Oscar.RayVector{Nemo.QQFieldElem}}:
  [0, 0, 1, -1]
  [0, 1, -1, 0]
  [1, -1, 0, 0]
 ```
 """
-@memoize Dict function sst(Q, e)
-  e_perp = Oscar.polyhedron([e, -e], [0, 0]) #e^{\perp}
-  all_gen = filter(
-    eprime -> !all(ei == 0 for ei in eprime) && eprime != e,
-    all_general_subdimension_vectors(Q, e),
-  )
-  isempty(all_gen) && return e_perp
-  return intersect(e_perp, Oscar.polyhedron(all_gen, zeros(Int, length(all_gen))))
-end
+function sst end
 
 """
     vgit_walls(Q, d; inner=false, top_dimension=true)
@@ -131,6 +84,8 @@ the `top_dimension=false` keyword to compute all the `W_e` of the VGIT problem.
 Defaults to computing all walls, pass the `inner=true` keyword
 to not include the outer walls.
 
+Requires Oscar: run `import Oscar` to enable this function.
+
 # Example
 
 ```jldoctests
@@ -138,7 +93,7 @@ julia> Q = Quiver("1-2,2-3,3-4,1-3,1-4"); d = [1, 1, 1, 1];
 
 julia> walls = vgit_walls(Q, d; inner=false, top_dimension=false);
 
-julia> map(QuiverTools.Oscar.dim, walls)
+julia> map(Oscar.dim, walls)
 7-element Vector{Int64}:
  2
  2
@@ -149,18 +104,7 @@ julia> map(QuiverTools.Oscar.dim, walls)
  2
 ```
 """
-@memoize Dict function vgit_walls(Q, d; inner=false, top_dimension=true)
-  all_subd = QuiverTools.all_subdimension_vectors(d; nonzero=true, strict=true)
-  out = map(
-    e -> reduce(intersect, [sst(Q, e), sst(Q, d - e), sst(Q, d)]), # definition of W_{e}
-    all_subd)
-  top_dim = maximum(Oscar.dim(w) for w in out) # max dimension of WALLS
-  top_dimension && filter!(w -> Oscar.dim(w) == top_dim, out)
-  inner && filter!(
-    w -> !any(issubset(w, f) for f in Oscar.facets(Oscar.Polyhedron, sst(Q, d))), out
-  )
-  return unique!(__helper_accelerate, out)
-end
+function vgit_walls end
 
 """
     wall_system(Q, d; inner=false, as_cones=true)
@@ -171,28 +115,9 @@ vgit wall `W_e` that lays on `H_e`.
 By default it returns the walls intersected with the semistable cone `sst(d)`,
 pass the keyword `as_cones=false` to get the full hyperplanes.
 
+Requires Oscar: run `import Oscar` to enable this function.
 """
-function wall_system(Q, d; inner=false, as_cones=true)
-  sstd = sst(Q, d)
-
-  all_walls = vgit_walls(Q, d; top_dimension=true)
-  if inner
-    all_walls = filter(
-      w -> !any(issubset(w, f) for f in Oscar.facets(Oscar.Polyhedron, sstd)), all_walls
-    )
-  end
-  # each affine hull contains the defining hyperplanes that cut it out.
-  all_walls = map(w -> Oscar.affine_hull(w), all_walls)
-  all_walls = map(
-    AH -> Oscar.polyhedron(
-      vcat([hyp.a[1, :] for hyp in AH], [-hyp.a[1, :] for hyp in AH]),
-      zeros(Int, 2 * length(AH)),
-    ),
-    all_walls,
-  )
-  as_cones && return map(w -> intersect(w, sstd), all_walls)
-  return all_walls
-end
+function wall_system end
 
 """
     vgit_chambers(Q, d; verbose=false)
@@ -207,6 +132,8 @@ and are defined by the walls `W_e` of codimension 1.
 If there exit no stable representations,
 `Oscar.dim(sst(Q, d))` is strictly smaller than `length(d) - 1`.
 
+Requires Oscar: run `import Oscar` to enable this function.
+
 # Example
 
 The following example has three inner walls `W_e`, and all of them are strictly smaller
@@ -215,7 +142,7 @@ than the corresponding `H_e \\cap sst(d)` in the wall system.
 ```jldoctests
 julia> Q = Quiver("1-2,2-3,3-4,1-4"); d = [1, 1, 1, 1];
 
-julia> map(QuiverTools.Oscar.rays, vgit_chambers(Q, d; verbose=false))
+julia> map(Oscar.rays, vgit_chambers(Q, d; verbose=false))
 3-element Vector{Oscar.SubObjectIterator{Oscar.RayVector{Nemo.QQFieldElem}}}:
  [[1, -1, 0, 0], [0, 0, 1, -1], [1, 0, 0, -1]]
  [[0, 1, -1, 0], [0, 0, 1, -1], [1, 0, 0, -1]]
@@ -228,7 +155,7 @@ hyperplane and the two others are not.
 ```jldoctests
 julia> Q = Quiver("1-2,2-3,3-4,1-3,1-4"); d = [1, 1, 1, 1];
 
-julia> map(QuiverTools.Oscar.rays, vgit_chambers(Q, d; verbose=false))
+julia> map(Oscar.rays, vgit_chambers(Q, d; verbose=false))
 4-element Vector{Oscar.SubObjectIterator{Oscar.RayVector{Nemo.QQFieldElem}}}:
  [[1, -1, 0, 0], [1, 0, 0, -1], [1, 0, -1, 0]]
  [[1, -1, 0, 0], [0, 0, 1, -1], [1, 0, 0, -1]]
@@ -236,152 +163,21 @@ julia> map(QuiverTools.Oscar.rays, vgit_chambers(Q, d; verbose=false))
  [[0, 1, -1, 0], [0, 0, 1, -1], [1, 0, 0, -1]]
 ```
 """
-@memoize Dict function vgit_chambers(Q, d; verbose=false)
-  sstd = __helper_accelerate(sst(Q, d))
-
-  sstd_dim = Oscar.dim(sstd)
-  # top-dimensional inner walls
-  int_walls = vgit_walls(Q, d; top_dimension=true)
-  int_walls = filter(
-    w -> !any(issubset(w, f) for f in Oscar.facets(Oscar.Polyhedron, sstd)), int_walls
-  )
-
-  # @warn "we must treat the case of a wall W_e of dimension = Oscar.dim(sstd) here!"
-
-  # we split the walls into two sets: the ones equal to the wall system hyperplane
-  # they lay on, and the ones that are not.
-  wallsyst = wall_system(Q, d; inner=true, as_cones=true)
-
-  full_walls = filter(w -> w in wallsyst, int_walls)
-  smaller_walls = filter(w -> !(w in wallsyst), int_walls)
-
-  full_walls = map(w -> Oscar.affine_hull(w), full_walls)
-  smaller_walls_ah = map(w -> Oscar.affine_hull(w), smaller_walls)
-
-  # find a vector not equal to d that cuts out the wall
-  full_walls = map(AH -> AH[findfirst(hyp -> (hyp.a[1, :] != d), AH)].a[1, :], full_walls)
-  smaller_walls_ah = map(
-    AH -> AH[findfirst(hyp -> (hyp.a[1, :] != d), AH)].a[1, :], smaller_walls_ah
-  )
-
-  full_walls_iterate = eachindex(full_walls)
-  smaller_walls_iterate = eachindex(smaller_walls_ah)
-
-  top_chambers = [sstd]
-
-  # helper function
-  function helper_split(wall, chamber)
-    return [
-      intersect(chamber, Oscar.polyhedron([wall], [0])),
-      intersect(chamber, Oscar.polyhedron([-wall], [0])),
-    ]
-  end
-  verbose &&
-    @info "There are $(length(full_walls)) full walls and $(length(smaller_walls)) smaller walls."
-
-  verbose && @info "Treating the full walls..."
-  for i in full_walls_iterate
-    if verbose
-      @time begin
-        # in-place
-        n = length(top_chambers)
-        for j in 1:n
-          push!(top_chambers,
-            helper_split(full_walls[i], top_chambers[j])...,
-          )
-        end
-        deleteat!(top_chambers, 1:n)
-
-        map!(__helper_accelerate, top_chambers, top_chambers)
-        filter!(c -> Oscar.dim(c) == sstd_dim, top_chambers)
-      end
-    else
-      # in-place
-      n = length(top_chambers)
-      for j in 1:n
-        push!(top_chambers,
-          helper_split(full_walls[i], top_chambers[j])...,
-        )
-      end
-      deleteat!(top_chambers, 1:n)
-
-      map!(__helper_accelerate, top_chambers, top_chambers)
-      filter!(c -> Oscar.dim(c) == sstd_dim, top_chambers)
-    end
-    verbose && @info "Found $(length(top_chambers)) unique chambers after $(i) steps.\n"
-  end
-
-  length(smaller_walls) == 0 && return top_chambers
-
-  verbose && @info "Treating the smaller walls..."
-  for i in smaller_walls_iterate
-    if verbose
-      @time begin
-        top_chambers = vcat(
-          map(
-            chamber -> if issubset(smaller_walls[i], chamber)
-              helper_split(smaller_walls_ah[i], chamber)
-            else
-              [chamber]
-            end,
-            top_chambers,
-          )...,
-        )
-        map!(__helper_accelerate, top_chambers, top_chambers)
-        filter!(c -> Oscar.dim(c) == length(d) - 1, top_chambers)
-      end
-    else
-      top_chambers = vcat(
-        map(
-          chamber -> if issubset(smaller_walls[i], chamber)
-            helper_split(smaller_walls_ah[i], chamber)
-          else
-            [chamber]
-          end,
-          top_chambers,
-        )...,
-      )
-      map!(__helper_accelerate, top_chambers, top_chambers)
-      filter!(c -> Oscar.dim(c) == sstd_dim, top_chambers)
-    end
-    verbose && @info "Found $(length(top_chambers)) unique chambers after $(i) steps.\n"
-  end
-
-  verbose && @info "Removing fake walls..."
-
-  # remove fake walls that appeared during the previous loop
-  incomplete = true
-  while incomplete
-    incomplete = false
-    for (ch1, ch2) in IterTools.subsets(top_chambers, 2)
-      inters = intersect(ch1, ch2)
-      # we are looking for common facets
-      Oscar.dim(inters) != sstd_dim - 1 && continue
-      # if their intersection (the common facet) lays on a W_e, good.
-      any(issubset(inters, w) for w in int_walls) && continue
-      verbose && @info "Found a fake wall, removing it..."
-      incomplete = true
-      # otherwise, we replace the two chambers in top_chambers by their union
-      push!(top_chambers, Oscar.minkowski_sum(ch1, ch2))
-      top_chambers = filter(ch -> (ch != ch1 && ch != ch2), top_chambers)
-    end
-  end
-  verbose &&
-    @info "Done. Found $(length(top_chambers)) chambers after removing the fake walls."
-  return top_chambers
-end
+function vgit_chambers end
 
 """
     vgit_fan(Q, d; verbose=false)
 
 Compute the VGIT fan for the quiver `Q` with dimension vector `d`.
 
+Requires Oscar: run `import Oscar` to enable this function.
+
 # Example
 
 ```jldoctests
 julia> Q = three_vertex_quiver(2, 3, 4); d = [2, 3, 4];
 
-julia> F = vgit_fan(Q, d); QuiverTools.Oscar.rays(F)
+julia> F = vgit_fan(Q, d); Oscar.rays(F)
 9-element Oscar.SubObjectIterator{Oscar.RayVector{Nemo.QQFieldElem}}:
  [1, -2//3, 0]
  [1, -1//2, -1//8]
@@ -394,13 +190,7 @@ julia> F = vgit_fan(Q, d); QuiverTools.Oscar.rays(F)
  [0, 1, -3//4]
 ```
 """
-function vgit_fan(Q, d; verbose=false)
-  return Oscar.polyhedral_fan(
-    map(ch -> Oscar.positive_hull(Oscar.rays(ch)),
-      vgit_chambers(Q, d; verbose=verbose),
-    ),
-  )
-end
+function vgit_fan end
 
 """
     git_equivalent(Q, d, theta1, theta2)
@@ -411,12 +201,14 @@ By [[Corollary 4.4, MR5007902](https://mathscinet.ams.org/mathscinet-getitem?mr=
 this is equivalent to their convex hull
 either lying in a wall or not intersecting any of them.
 
+Requires Oscar: run `import Oscar` to enable this function.
+
 # Example
 
 ```jldoctests
 julia> Q = three_vertex_quiver(2, 3, 4); d = [1, 2, 2];
 
-julia> F = vgit_fan(Q, d); QuiverTools.Oscar.rays(F)
+julia> F = vgit_fan(Q, d); Oscar.rays(F)
 4-element Oscar.SubObjectIterator{Oscar.RayVector{Nemo.QQFieldElem}}:
  [1, -1//2, 0]
  [1, 0, -1//2]
@@ -459,54 +251,19 @@ julia> any(y in w for w in W)
 true
 ```
 """
-function git_equivalent(Q, d, theta1, theta2)
-  theta1 == theta2 && return true
-  line = Oscar.convex_hull([theta1, theta2]) # 1-dimensional iif theta1 != theta2
+function git_equivalent end
 
-  # either the line lies in a wall or it intersects none of them
-  for w in vgit_walls(Q, d; top_dimension=false)
-    if !Oscar.issubset(line, w) && Oscar.is_feasible(Oscar.intersect(line, w))
-      return false
-    end
-  end
-  return true
-end
-
-"""
-    __sst_cone(Q::Quiver, e::AbstractVector{Int})
-
-Return the exact same polyhedral cone as `sst(Q, e)`
-but seen as an Oscar `Cone` object.
-
-For internal use only for now.
-"""
-function __sst_cone(Q::Quiver, e::AbstractVector{Int})
-  all_gen = filter(
-    eprime -> !all(ei == 0 for ei in eprime) && eprime != e,
-    all_general_subdimension_vectors(Q, e),
+# When Oscar is not loaded, every entry point above resolves to this catch-all,
+# which explains how to enable it. The extension adds concrete methods that take
+# precedence once Oscar has been loaded.
+for f in (
+  :is_special_subdimension_vector, :all_special_subdimension_vectors, :sst,
+  :vgit_walls, :wall_system, :vgit_chambers, :vgit_fan, :git_equivalent,
+)
+  @eval $f(args...; kwargs...) = throw(
+    ArgumentError(
+      "`" * $(string(f)) * "` needs Oscar's polyhedral geometry; run `import Oscar` " *
+      "to enable the walls-and-chambers / VGIT functions.",
+    ),
   )
-  isempty(all_gen) && return Oscar.cone_from_inequalities([e, -e])
-  return Oscar.cone_from_inequalities(all_gen, [e])
-end
-
-"""
-    __lower_fan(Q::Quiver, d::AbstractVector{Int})
-
-Compute the polyhedral fan from the definition
-using the fan construction via cones.
-
-Does not contain the top dimensional chambers,
-so it cannot be used to iterate over chambers.
-
-It is way faster and more likely to be correct in G.P. opinion.
-
-For internal use only for now.
-"""
-function __lower_fan(Q::Quiver, d::AbstractVector{Int})
-  walls = map(
-    e -> Oscar.intersect(__sst_cone(Q, e), __sst_cone(Q, d - e)),
-    QuiverTools.all_subdimension_vectors(d; nonzero=true, strict=true),
-  )
-  isempty(walls) && return Oscar.polyhedral_fan(__sst_cone(Q, d)) # I guess
-  return Oscar.polyhedral_fan(walls)
 end
